+++ src/main/java/com/example/puddingbe/domain/timer/presentation/dto/TimerResponse.java
@@ -1,4 +1,4 @@
-package com.example.puddingbe.domain.timer.domain.dto;
+package com.example.puddingbe.domain.timer.presentation.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
