+++ src/main/java/com/example/puddingbe/domain/user/presentation/dto/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.example.puddingbe.domain.user.request;
+package com.example.puddingbe.domain.user.presentation.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
