+++ src/main/java/com/example/puddingbe/global/detail/DetailService.java
@@ -1,7 +1,7 @@
 package com.example.puddingbe.global.detail;
 
 import com.example.puddingbe.domain.user.domain.User;
-import com.example.puddingbe.domain.user.repository.UserRepository;
+import com.example.puddingbe.domain.user.domain.repository.UserRepository;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
