+++ src/main/java/com/example/puddingbe/global/detail/UserFacade.java
@@ -1,14 +1,11 @@
 package com.example.puddingbe.global.detail;
 
-import com.example.puddingbe.domain.user.domain.User;
-
-import com.example.puddingbe.domain.user.repository.UserRepository;
+import com.example.puddingbe.domain.user.domain.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
 import org.springframework.web.server.ResponseStatusException;
-import java.util.Optional;
 
 @Component
 @RequiredArgsConstructor
